Refactor Libraries to libraries
