+++ apps/src/code-studio/components/ShareAllowedDialog.jsx
@@ -24,7 +24,7 @@ import {createHiddenPrintWindow} from '@cdo/apps/utils';
 import i18n from '@cdo/locale';
 import firehoseClient from '@cdo/apps/lib/util/firehose';
 import experiments from '@cdo/apps/util/experiments';
-import LibraryCreationDialog from './Libraries/LibraryCreationDialog';
+import LibraryCreationDialog from './libraries/LibraryCreationDialog';
 
 function recordShare(type) {
   if (!window.dashboard) {
